+++ src/connector.ts
@@ -1,6 +1,6 @@
 import { Registry } from "prom-client";
 import {
-  CapabilitiesResponse,
+  Capabilities,
   QueryRequest,
   QueryResponse,
   SchemaResponse,
@@ -35,6 +35,7 @@ export interface Connector<Configuration, State> {
     configuration: Configuration,
     metrics: Registry
   ): Promise<State>;
+
   /**
    *
    * Update any metrics from the state
@@ -48,17 +49,21 @@ export interface Connector<Configuration, State> {
    * @param state
    */
   fetchMetrics(configuration: Configuration, state: State): Promise<undefined>;
+
   /**
    * Check the health of the connector.
    *
-   * For example, this function should check that the connector
-   * is able to reach its data source over the network.
+   * This should simply verify that the connector is ready to start accepting
+   * requests. It should not verify that external data sources are available.
    *
-   * Should throw if the check fails, else resolve
+   * This function is optional to implement; if left unimplemented, a default
+   * implementation will be used that returns healthy once the connector
+   * webserver is running.
+   *
    * @param configuration
    * @param state
    */
-  healthCheck(configuration: Configuration, state: State): Promise<undefined>;
+  getHealthReadiness?(configuration: Configuration, state: State): Promise<undefined>;
 
   /**
    * Get the connector's capabilities.
@@ -66,10 +71,10 @@ export interface Connector<Configuration, State> {
    * This function implements the [capabilities endpoint](https://hasura.github.io/ndc-spec/specification/capabilities.html)
    * from the NDC specification.
    *
-   * This function should be syncronous
+   * This function should be synchronous
    * @param configuration
    */
-  getCapabilities(configuration: Configuration): CapabilitiesResponse;
+  getCapabilities(configuration: Configuration): Capabilities;
 
   /**
    * Get the connector's schema.